code/trivial_baselines/requirements.txt
numpy==1.21.2
faiss-cpu==1.7.1
faiss-gpu==1.6.3